meaningful change 2
A Youtube video about music visualization gave me an inspiration. So I wrote it in ps.j5 to figure out how he did it? And what code he used and why? This change is just adding the code from the video so I didn't come up with most of the code. Here's the link to the video: https://www.youtube.com/watch?v=8B3GzZo-z9A